fix no.96

+++ asafov.andrew/T1/outputFunctions.hpp
@@ -9,7 +9,7 @@ namespace asafov
 {
   std::string getName(std::istream& in);
   Shape* ShapeFactory(std::string shapename, std::istream& in);
-  void scaleShapes(Shape** shapes, unsigned long long count, point_t pos, double scale, std::ostream& out);
+  void scaleShapes(Shape** shapes, size_t count, point_t pos, double scale, std::ostream& out);
   void deleteShapes(Shape** shapes, size_t count);
 }
 #endif
